Add files via upload
//opening file
Opening the File:
The file is opened using fstream with the following flags to open it for reading and writing:

ios::in: Opens for reading.
ios::out: Opens for writing.
ios::app: New data will be appended to the end and not overwrite any existing content.
Checking for Errors
The program checks if the file opened successfully. It terminates if the file could not open by printing an error.

Reading the File:
It reads in and prints each line of the file using getline.

Writing to the File:
Before write it simply clears any read-related flags and moves the write pointer to the end of the file (seekp), then appends new data.

Closing the File:
After the operations, the file is closed to free resources.