Add files via upload
//texting file

This program demonstrates how to use C++ to write your data to a text file. This program achieves the following:

Opens a file called "manu.txt.".
Check whether this file opened successfully.
Writes two lines of text to the file.
Closes the file after writing.
This will print a success message to the console if it succeeds, or an error message if the file cannot be opened.