Add files via upload
//reading from a text file

This program reads the contents of a file-called manu.txt-and prints each line to the console. It :

Tries to open the file "manu.txt" with an ifstream object.
If the file opens correctly, it reads the file line by line using getline(), and prints to the console every line.
After reading all lines, it closes the file.
If it can't open the file, it just prints an error message.