Add files via upload
//binary file handling

This program shows the following things:

How to declare and use a structure (Data) to store data.
How to output binary data to a file using ofstream with ios::binary.
Use reinterpret_cast to write structured data-a struct-to a binary file as raw bytes.
Handle file errors - e.g. if file can't be opened.
print a success message to the console after having written the data.      

Example of contents of data.bin:
What the contents of data.bin will be: Data.bin won't be human-readable since it is in binary. It will contain the raw memory representation of the Data structure, namely a composition of:

4 bytes for the integer id (in little-endian or big-endian format depending on the system).
50 bytes for the name array, with the string "Mansi" followed by padding because of the array's 50 characters.